Let "ghosts" provide more probing options.
It is now possible to query entries, macros, tags and also have a
summary report of all tags and macros along with the amount of hosts
they are selecting.

To avoid mistakes, the parsing of the ghosts file is now validating that
we are not duplicating tags, mixing tags with hosts, etc.