gsh: use significant exit codes.
In particular, when a command exits with an error status on one host, we
need to know.  Also when the "gsh" command is interrupted, something bad
may happen remotely (especially when using "gsh" to remotely update a
system), hence we report an error status as well.

Being sure that all commands ran successfully is important when using
"gsh" in scripts.  Unfortunately, it is not easy to determine which
hosts failed, unless we parse the error output.